Created new TalonSRX class
Changed naming to reflect changes in CTRE
Still needs passthrough methods
